+++ src/main/java/com/example/libreria/service/GeneralService.java
@@ -7,9 +7,9 @@ import java.util.List;
 @Service
 public interface GeneralService<E> {
     public List<E> searchAll() throws Exception;
-    public E searchById(Integer id) throws Exception;
+    public E searchById(Long id) throws Exception;
     public E create(E data) throws Exception;
-    public E update(Integer id, E data) throws Exception;
-    public boolean delete(Integer id) throws Exception;
+    public E update(Long id, E data) throws Exception;
+    public boolean delete(Long id) throws Exception;
 
 }
